README fixes and refactors

+++ Console/Command/CacheFlush.php
@@ -16,6 +16,8 @@ class CacheFlush extends Command
 {
     public const TRANSLATION = 'translation';
 
+    public const LIST = 'list';
+
     /**
      * @var array
      */
@@ -51,7 +53,8 @@ class CacheFlush extends Command
     {
         $this->setName('torah:cache:flush');
         $this->setDescription('JaroslawZielinski Torah cache flush');
-        $this->addArgument(self::TRANSLATION, InputArgument::REQUIRED, (string)__('translation'));
+        $this->addOption(self::LIST,'l', $mode = null, (string)__('see translation codes avaiable.'));
+        $this->addArgument(self::TRANSLATION, InputArgument::OPTIONAL, (string)__('translation code'));
     }
 
     private function displayMessages(OutputInterface $output): int
@@ -67,38 +70,35 @@ class CacheFlush extends Command
         $this->messages[] = sprintf($message, ...$args);
     }
 
+    private function listTranslations(OutputInterface $output): void
+    {
+        foreach (Resources::TORAH_TRANSLATIONS as $code => $label) {
+            $output->writeln(
+                (string)__('%1 <fg=yellow;options=bold>%2</>', $code, str_replace(['&apos;'], ['\''], $label))
+            );
+        }
+    }
+
     /**
      * {@inheritDoc}
      * @throws \Exception
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        try {
-            $translation = $input->getArgument(self::TRANSLATION);
+        $translation = $input->getArgument(self::TRANSLATION) ?? null;
+        $list = $input->getOption(self::LIST);
+        if ($list) {
+            $this->listTranslations($output);
+        } else if ($translation && !$list) {
             $availableTranslations = array_keys(Resources::TORAH_TRANSLATIONS);
             if (!in_array($translation, $availableTranslations)) {
-                $availableTranslationsStr = print_r($availableTranslations, true);
-                $message = str_replace(['Array
-(
-', '
-)
-'], ['', ''], $availableTranslationsStr);
-            throw new \Exception(sprintf((string)__('Use the following available translations:
-%s'), $message));
+                throw new \Exception((string)__('Unkown translation: %1', $translation));
             }
             $result = $this->repository->delete($translation);
             $resultMsg = sprintf((string)__('For \'%s\' cache flushed with result \'%s\''), $translation, $result);
             $output->writeln($resultMsg);
-        } catch (\Exception $e) {
-                       $message = sprintf(
-                '<fg=red;options=bold>Something went wrong</>: <fg=white>\'%s\'</>: <fg=yellow>
-%s</>.',
-                $e->getMessage(),
-                $e->getTraceAsString()
-            );
-            $this->logger->error($e->getMessage(), $e->getTrace());
-            $output->writeln($message);
-            return 1;
+        } else {
+            throw new \Exception(sprintf((string)__('Translation parameter empty.')));
         }
         return $this->displayMessages($output);
     }
